Add ForwardAsMime setting to control the default forwarding behavior
MailFlow forwards messages as MIME attachments by default and inline if the
Option key is held down. Allow changing the default behavior by setting the
option ForwardAsMime. The default is on. Setting it to false causes messages to
be forwarded inline by default and as MIME attachments if the Option key is held
down.